facial/requirements.txt
fastapi
uvicorn
python-dotenv
face_recognition
numpy
httpx